test(ui): extract makeFakeStdin/Stdout to shared tests/helpers/ink-stdio
